Add files via upload
A buyTicket function that queries the user for their age (under 12 or over 65 gets a discount on a ticket), and then queries the user for yes or no response to "Are you attending a matinee show?". Answering yes or y or Y will deduct $3 from the price, likewise answering a version of "no" will not. An alert box will display the final price for the ticket.